refactor: move `fibonacci_series` function to `series.py` file
