+++ lib/starter/__init__.py
@@ -1,5 +1,3 @@
-def fibonacci_series(n: int) -> list[int]:
-    series = [0, 1]
-    for i in range(n - 1):
-        series.append(series[-1] + series[-2])
-    return series[1:]
+from .series import fibonacci_series
+
+__all__ = ["fibonacci_series"]
